+++ Assets/_Scripts/Interfaces/ICharacter.cs
@@ -1,6 +1,6 @@
 public interface ICharacter
 {
-    void TakeDamage ( int damageAmount, int shooterIndex );
+    void TakeDamage ( float damageAmount, int shooterIndex );
     void Die (int killerIndex);
 
     void Freeze ( float duration );
